requirements.txt
Flask==1.0.1
flask_sqlalchemy==2.3.1
flask_wtf==0.14.2
Werkzeug==0.14.1
WTForms==2.1
